+++ src/ZivoeRewardsVesting.sol
@@ -137,7 +137,7 @@ contract ZivoeRewardsVesting is ReentrancyGuard, Ownable {
     /// @notice This modifier ensures user rewards information is updated BEFORE mutative actions.
     /// @param account The account to update personal rewards information of (if not address(0)).
     modifier updateReward(address account) {
-        for (uint i; i < rewardTokens.length; i++) {
+        for (uint256 i; i < rewardTokens.length; i++) {
             address token = rewardTokens[i];
             rewardData[token].rewardPerTokenStored = rewardPerToken(token);
             rewardData[token].lastUpdateTime = lastTimeRewardApplicable(token);
@@ -377,7 +377,7 @@ contract ZivoeRewardsVesting is ReentrancyGuard, Ownable {
 
     /// @notice Claim rewards for all possible _rewardTokens.
     function getRewards() public nonReentrant updateReward(_msgSender()) {
-        for (uint i; i < rewardTokens.length; i++) { getRewardAt(i); }
+        for (uint256 i; i < rewardTokens.length; i++) { getRewardAt(i); }
     }
     
     /// @notice Claim rewards for a specific _rewardToken.
