+++ src/lockers/OCL/OCL_ZVE_UNIV2.sol
@@ -108,7 +108,7 @@ contract OCL_ZVE_UNIV2 is ZivoeLocker, ZivoeSwapper {
             "OCL_ZVE_UNIV2::pushToLockerMulti() assets[0] != pairAsset || assets[1] != IZivoeGlobals(GBL).ZVE()"
         );
 
-        for (uint i = 0; i < 2; i++) {
+        for (uint256 i = 0; i < 2; i++) {
             require(amounts[i] >= 10 * 10**6, "OCL_ZVE_UNIV2::pushToLockerMulti() amounts[i] < 10 * 10**6");
             IERC20(assets[i]).safeTransferFrom(owner(), address(this), amounts[i]);
         }
